use inject_trigger=any where unsure (+ suggestions to modify style format)
line 318 of hordelib/horde.py, if inject_trigger==any, it uses the first trigger. Since this default is *any*, it would probably be better to pick one at random.

suggestions:

It would also be nice if this could be an array rather than a string, so it could pick a trigger at random when submitting, it would make the bot more entertaining if we could get a random-ish style from multi-concept loras
On that note, amy of the following could also make the bot more entertaining:
- config item to pick a random lora from the list (all vs any vs current any=first)
- picking a random lora could also be useful
- allow use of categories in the bot (I'll suggest this to zelda_fan), when someone uses a category as a style it picks a style from that category randomly for each generation