Add more punk styles, some horror styles, and some glitch styles (#7)
* add glitch art styles

* add more stylepunks

* add horrors

* remove duplicate style

I had only searched for the Lora name with quotes around it as I'd noticed the first bunch were banned as such and assumed they were all like that. If I hadn't had the quotes, I would have caught this.

* Add bodyhorror, this one's pretty disgusting

I've added all lora fields since none of the others are around to show how they work without tracking it down in the api reference. I also included the trigger_inject because theres no documentation on what this does (I had to track it down in hordelib)

* Add ivorygoldai

* add nastyhorrors, combination of reelhorror and bodyhorror

* use inject_trigger=any where unsure (+ suggestions to modify style format)

line 318 of hordelib/horde.py, if inject_trigger==any, it uses the first trigger. Since this default is *any*, it would probably be better to pick one at random.

suggestions:

It would also be nice if this could be an array rather than a string, so it could pick a trigger at random when submitting, it would make the bot more entertaining if we could get a random-ish style from multi-concept loras
On that note, amy of the following could also make the bot more entertaining:
- config item to pick a random lora from the list (all vs any vs current any=first)
- picking a random lora could also be useful
- allow use of categories in the bot (I'll suggest this to zelda_fan), when someone uses a category as a style it picks a style from that category randomly for each generation

* add sdxl styles

* fix json mistakes

* add my own horror prompt 👍

whoops. accidentally commited after running dos2unix (git diff --ignore-cr-at-eol isnt useful for copypaste...)

* add sdxl chibis